+++ src/forze/application/contracts/document/__init__.py
@@ -5,6 +5,7 @@ from .deps import (
     DocumentQueryDepKey,
     DocumentQueryDepPort,
 )
+from .gateways import DocumentReadGatewayPort, DocumentWriteGatewayPort
 from .helpers import require_create_id, require_create_id_for_many
 from .ports import DocumentCommandPort, DocumentQueryPort
 from .specs import DocumentSpec, DocumentWriteTypes
@@ -13,6 +14,8 @@ from .types import RowLockMode
 # ----------------------- #
 
 __all__ = [
+    "DocumentReadGatewayPort",
+    "DocumentWriteGatewayPort",
     "DocumentQueryPort",
     "DocumentCommandPort",
     "RowLockMode",